test: move spec to frontend-new/specs/ where playwright globs it
Etherpad core's playwright config picks up plugin specs at
`static/tests/frontend-new/specs/**/*.spec.ts` (not `frontend/specs/**/*.spec.js`).
The test was being silently skipped by CI. Also rewrote to use the standard
goToNewPad/goToPad helpers and TypeScript syntax matching the existing
cursortrace.spec.ts in this directory.
